Transcript

+++ src/script.js
@@ -5,6 +5,8 @@
 /**
  *
  */
+var what = process.argv.slice(2)[0];
+
 if(process.env.CIRCLE_BRANCH === 'develop'){
     var SSH_USERNAME = process.env.DEV_SSH_USERNAME;
     var SSH_PASSWORD = process.env.DEV_SSH_PASSWORD;
@@ -18,34 +20,46 @@ if(process.env.CIRCLE_BRANCH === 'develop'){
         user: SSH_USERNAME,
         pass: SSH_PASSWORD
     });
-    ssh
-        .exec('cd /home/cbo/docker/api', {
-            out: console.log.bind(console)
-        })
-        .exec('git pull origin develop', {
-            out: console.log.bind(console)
-        })
-        .exec('docker rm -f api', {
-            out: console.log.bind(console)
-        })
-        .exec('docker build -t psesd/ssl-api:develop .', {
-            out: console.log.bind(console)
-        })
-        .exec('docker run -d --name api --link redis:redis -p 104.192.103.12:443:443 -e NODE_ENV=development -e NODE_CONFIG_DIR=/config -v /config:/config psesd/ssl-api:develop', {
-            out: console.log.bind(console)
-        })
-        .exec('docker login -e "'+DOCKER_EMAIL+'" -u "'+DOCKER_USER+'" -p "'+DOCKER_PASS+'"', {
-            out: console.log.bind(console)
-        })
-        .exec('docker push psesd/ssl-api:develop', {
-            out: console.log.bind(console)
-        })
-        .exec('echo "DONE"', {
-            exit: function(code) {
-                process.exit(0);
-            }
-        })
-        .start();
+    if(what === 'build') {
+        ssh
+            .exec('cd /home/cbo/docker/api', {
+                out: console.log.bind(console)
+            })
+            .exec('git fetch && git reset --hard origin/develop', {
+                out: console.log.bind(console)
+            })
+            .exec('docker rm -f api', {
+                out: console.log.bind(console)
+            })
+            .exec('docker build -t psesd/ssl-api:develop .', {
+                out: console.log.bind(console)
+            })
+            .exec('docker run -d --name api --link redis:redis -p 104.192.103.12:443:443 -e NODE_ENV=development -e NODE_CONFIG_DIR=/config -v /config:/config psesd/ssl-api:develop', {
+                out: console.log.bind(console)
+            })
+            .exec('echo "DONE"', {
+                exit: function (code) {
+                    process.exit(0);
+                }
+            });
+    } else {
+        ssh
+            .exec('cd /home/cbo/docker/api', {
+                out: console.log.bind(console)
+            })
+            .exec('docker login -e "' + DOCKER_EMAIL + '" -u "' + DOCKER_USER + '" -p "' + DOCKER_PASS + '"', {
+                out: console.log.bind(console)
+            })
+            .exec('docker push psesd/ssl-api:develop', {
+                out: console.log.bind(console)
+            })
+            .exec('echo "DONE"', {
+                exit: function (code) {
+                    process.exit(0);
+                }
+            })
+    }
+    ssh.start();
 
 } else {
     throw 'Not valid environment';
